zshrc: top two directories
I've started using git worktrees more and I tend to do:
`$HOME/company/repo/branch`

So when I'm sitting in a repo, I'm actually in something like
`repo/main`. So the current prompt (prior to this change) is just `main`
which is not very helpful. But after this change, it becomes
`repo/main`.